requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
flask-session
flask-limiter
PyJWT>=2.0.0
bcrypt>=3.2.0
requests>=2.25.0
langchain>=0.3.4
langchain-community>=0.3.3
langchain-core>=0.3.47
langchain-openai>=0.3.3
langchain-ollama>=0.3.0
langgraph
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.45
pydantic==2.12.5
python-dotenv==1.0.1
openai>=1.65.2
tiktoken==0.8.0
gigachat>=0.1.41.post1,<0.2.0
langchain-gigachat==0.3.12
tenacity==8.5.0
chromadb==0.6.3
sentence-transformers==3.2.1
unstructured==0.16.25
pypdf==5.1.0
langchain-text-splitters>=0.3.5
langchain-huggingface==0.1.2
langchain-qdrant>=0.3.0
qdrant-client>=1.12.0
gunicorn>=22.0.0
redis>=5.2.0
streamlit>=1.28.0
# Note: The graphviz Python module requires the system-level Graphviz software to be installed
# On Debian/Ubuntu systems: sudo apt-get install graphviz
# On macOS: brew install graphviz
# On Windows: Install from https://graphviz.org/download/
graphviz
grandalf
marker-pdf>=0.1.0